Split the OS3 tests in the different test files.
