+++ tests/system/action/poll/test_vote.py
@@ -1,3 +1,4 @@
+from openslides_backend.models.models import Poll
 from tests.system.action.base import BaseActionTestCase
 
 
@@ -435,3 +436,1365 @@ class PollVoteTest(BaseActionTestCase):
         user = self.get_model("user/1")
         assert user.get("vote_$_ids") == ["113"]
         assert user.get("vote_$113_ids") == [1]
+
+
+class VotePollBaseTestClass(BaseActionTestCase):
+    def setUp(self) -> None:
+        super().setUp()
+        self.create_model(
+            "assignment/1",
+            dict(
+                title="test_assignment_tcLT59bmXrXif424Qw7K",
+                open_posts=1,
+                candidate_ids=[1],
+            ),
+        )
+        self.create_poll()
+        self.create_model("meeting/113", {"name": "my meeting"})
+        self.create_model("group/1", {"user_ids": [1]})
+        self.create_model("option/1", {"meeting_id": 113, "poll_id": 1})
+        self.create_model("option/2", {"meeting_id": 113, "poll_id": 1})
+        self.update_model(
+            "user/1",
+            {
+                "is_present_in_meeting_ids": [113],
+                "group_$113_ids": [1],
+                "group_$_ids": ["113"],
+            },
+        )
+
+    def create_poll(self) -> None:
+        # has to be implemented by subclasses
+        raise NotImplementedError()
+
+    def start_poll(self) -> None:
+        self.update_model("poll/1", {"state": Poll.STATE_STARTED})
+
+    def add_candidate(self) -> None:
+        self.create_model("option/3", {"meeting_id": 113, "poll_id": 1})
+        self.update_model("poll/1", {"option_ids": [1, 2, 3]})
+
+
+class VotePollOS3AnalogYNA(VotePollBaseTestClass):
+    def create_poll(self) -> None:
+        self.create_model(
+            "poll/1",
+            dict(
+                content_object_id="assignment/1",
+                title="test_title_04k0y4TwPLpJKaSvIGm1",
+                pollmethod="YNA",
+                type=Poll.TYPE_ANALOG,
+                state=Poll.STATE_CREATED,
+                meeting_id=113,
+                option_ids=[1, 2],
+                entitled_group_ids=[1],
+            ),
+        )
+
+    def test_vote(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {
+                "value": {
+                    "options": {
+                        "1": {"Y": "1", "N": "2.35", "A": "-1"},
+                        "2": {"Y": "30", "N": "-2", "A": "8.93"},
+                    },
+                    "votesvalid": "4.64",
+                    "votesinvalid": "-2",
+                    "votescast": "-2",
+                },
+                "user_id": 1,
+                "id": 1,
+            },
+        )
+        self.assert_status_code(response, 200)
+        self.assert_model_exists("vote/6")
+        self.assert_model_not_exists("vote/7")
+        poll = self.get_model("poll/1")
+        self.assertEqual(poll.get("votesvalid"), "4.64")
+        self.assertEqual(poll.get("votesinvalid"), "-2")
+        self.assertEqual(poll.get("votescast"), "-2")
+        self.assertEqual(poll.get("state"), Poll.STATE_FINISHED)
+        option1 = self.get_model("option/1")
+        option2 = self.get_model("option/2")
+        self.assertEqual(option1.get("yes"), "1")
+        self.assertEqual(option1.get("no"), "2.35")
+        self.assertEqual(option1.get("abstain"), "-1")
+        self.assertEqual(option2.get("yes"), "30")
+        self.assertEqual(option2.get("no"), "-2")
+        self.assertEqual(option2.get("abstain"), "8.93")
+
+    def test_vote_fractional_negative_values(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {
+                "value": {
+                    "options": {"1": {"Y": "1", "N": "1", "A": "1"}},
+                    "votesvalid": "-1.5",
+                    "votesinvalid": "-2",
+                    "user_id": 1,
+                    "id": 1,
+                    "votescast": "-2",
+                },
+            },
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_too_many_options(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {
+                "value": {
+                    "options": {
+                        "1": {"Y": "1", "N": "2.35", "A": "-1"},
+                        "2": {"Y": "1", "N": "2.35", "A": "-1"},
+                    }
+                },
+                "user_id": 1,
+                "id": 1,
+            },
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_too_few_options(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {
+                "value": {"options": {"1": {"Y": "1", "N": "2.35", "A": "-1"}}},
+                "user_id": 1,
+                "id": 1,
+            },
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_wrong_options(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {
+                "value": {
+                    "options": {
+                        "1": {"Y": "1", "N": "2.35", "A": "-1"},
+                        "3": {"Y": "1", "N": "2.35", "A": "-1"},
+                    }
+                },
+                "user_id": 1,
+                "id": 1,
+            },
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_wrong_state(self) -> None:
+        response = self.request("poll.vote", {"value": {}, "id": 1, "user_id": 1})
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_missing_value(self) -> None:
+        self.start_poll()
+        response = self.request("poll.vote", {"value": {}, "id": 1, "user_id": 1})
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_wrong_data_format(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote", {"value": [1, 2, 5], "id": 1, "user_id": 1}
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_wrong_option_format(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"options": [1, "string"]}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_wrong_option_id_type(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {
+                "value": {"options": {"string": "some_other_string"}},
+                "id": 1,
+                "user_id": 1,
+            },
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_wrong_vote_data(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"options": {"1": [None]}, "id": 1, "user_id": 1}},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_missing_vote_value(self) -> None:
+        self.start_poll()
+        for value in "YNA":
+            data = {
+                "options": {"1": {"Y": "1", "N": "3", "A": "-1"}},
+            }
+            del data["options"]["1"][value]
+            response = self.request("poll.vote", {"value": data, "id": 1, "user_id": 1})
+            self.assert_status_code(response, 400)
+            self.assert_model_not_exists("vote/1")
+
+    def test_vote_state_finished(self) -> None:
+        self.start_poll()
+        self.request(
+            "poll.vote",
+            {
+                "value": {
+                    "options": {"1": {"Y": 5, "N": 0, "A": 1}},
+                    "votesvalid": "-2",
+                    "votesinvalid": "1",
+                    "votescast": "-1",
+                },
+                "id": 1,
+                "user_id": 1,
+            },
+        )
+        self.update_model("poll/1", dict(state="finished"))
+        response = self.request(
+            "poll.vote",
+            {
+                "value": {
+                    "options": {"1": {"Y": 2, "N": 2, "A": 2}},
+                    "votesvalid": "4.64",
+                    "votesinvalid": "-2",
+                    "votescast": "3",
+                },
+                "id": 1,
+                "user_id": 1,
+            },
+        )
+        self.assert_status_code(response, 200)
+        poll = self.get_model("poll/1")
+        self.assertEqual(poll.get("votesvalid"), "4.64")
+        self.assertEqual(poll.get("votesinvalid"), "-2")
+        self.assertEqual(poll.get("votescast"), "3")
+        self.assert_model_exists("vote/3")
+        self.assert_model_not_exists("vote/4")
+        option = self.get_model("option/1")
+        self.assertEqual(option.get("yes"), "2")
+        self.assertEqual(option.get("no"), "2")
+        self.assertEqual(option.get("abstain"), "2")
+
+
+class VotePollOS3NamedYNA(VotePollBaseTestClass):
+    def create_poll(self) -> None:
+        self.create_model(
+            "poll/1",
+            dict(
+                content_object_id="assignment/1",
+                title="test_title_OkHAIvOSIcpFnCxbaL6v",
+                pollmethod="YNA",
+                type=Poll.TYPE_NAMED,
+                state=Poll.STATE_CREATED,
+                meeting_id=113,
+                option_ids=[1, 2],
+                entitled_group_ids=[1],
+            ),
+        )
+
+    def test_vote(self) -> None:
+        self.add_candidate()
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": "Y", "2": "N", "3": "A"}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 200)
+        self.assert_model_exists("vote/1")
+        self.assert_model_exists("vote/2")
+        self.assert_model_exists("vote/3")
+        self.assert_model_not_exists("vote/4")
+        poll = self.get_model("poll/1")
+        self.assertEqual(poll.get("votesvalid"), "1.000000")
+        self.assertEqual(poll.get("votesinvalid"), "0.000000")
+        self.assertEqual(poll.get("votescast"), "1.000000")
+        self.assertEqual(poll.get("state"), Poll.STATE_STARTED)
+        self.assertIn(1, poll.get("voted_ids", []))
+        option1 = self.get_model("option/1")
+        option2 = self.get_model("option/2")
+        option3 = self.get_model("option/3")
+        self.assertEqual(option1.get("yes"), "1.000000")
+        self.assertEqual(option1.get("no"), "0.000000")
+        self.assertEqual(option1.get("abstain"), "0.000000")
+        self.assertEqual(option2.get("yes"), "0.000000")
+        self.assertEqual(option2.get("no"), "1.000000")
+        self.assertEqual(option2.get("abstain"), "0.000000")
+        self.assertEqual(option3.get("yes"), "0.000000")
+        self.assertEqual(option3.get("no"), "0.000000")
+        self.assertEqual(option3.get("abstain"), "1.000000")
+
+    def test_vote_with_voteweight(self) -> None:
+        # config["users_activate_vote_weight"] = True
+        self.update_model("user/1", dict(vote_weight="4.200000"))
+        self.add_candidate()
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": "Y", "2": "N", "3": "A"}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 200)
+        self.assert_model_exists("vote/1")
+        self.assert_model_exists("vote/2")
+        self.assert_model_exists("vote/3")
+        self.assert_model_not_exists("vote/4")
+        poll = self.get_model("poll/1")
+        self.assertEqual(poll.get("votesvalid"), "4.200000")
+        self.assertEqual(poll.get("votesinvalid"), "0.000000")
+        self.assertEqual(poll.get("votescast"), "1.000000")
+        self.assertEqual(poll.get("state"), Poll.STATE_STARTED)
+        option1 = self.get_model("option/1")
+        option2 = self.get_model("option/2")
+        option3 = self.get_model("option/3")
+        self.assertEqual(option1.get("yes"), "4.200000")
+        self.assertEqual(option1.get("no"), "0")
+        self.assertEqual(option1.get("abstain"), "0")
+        self.assertEqual(option2.get("yes"), "0")
+        self.assertEqual(option2.get("no"), "4.200000")
+        self.assertEqual(option2.get("abstain"), "0")
+        self.assertEqual(option3.get("yes"), "0")
+        self.assertEqual(option3.get("no"), "0")
+        self.assertEqual(option3.get("abstain"), "4.200000")
+
+    def test_change_vote(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": "Y"}, "id": 1, "user_id": 1},
+        )
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": "N"}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/2")
+        vote = self.get_model("vote/1")
+        self.assertEqual(vote.get("value"), "Y")
+
+    def test_too_many_options(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": "Y", "2": "N"}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_wrong_options(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": "Y", "3": "N"}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_no_permissions(self) -> None:
+        self.start_poll()
+        # self.make_admin_delegate()
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": "Y"}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_anonymous(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": "Y"}, "id": 1, "user_id": 0},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_vote_not_present(self) -> None:
+        self.start_poll()
+        self.update_model("user/1", dict(is_present_in_meeting_ids=[]))
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": "Y"}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_wrong_state(self) -> None:
+        response = self.request("poll.vote", {"value": {}, "id": 1, "user_id": 1})
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_missing_data(self) -> None:
+        self.start_poll()
+        response = self.request("poll.vote", {"value": {}, "id": 1, "user_id": 1})
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+        poll = self.get_model("poll/1")
+        self.assertNotIn(1, poll.get("voted_ids", []))
+
+    def test_wrong_data_format(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": [1, 2, 5], "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_wrong_option_format(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": "string"}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_wrong_option_id_type(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"id": "Y"}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_wrong_vote_data(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": [None]}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+
+class VotePollOS3NamedY(VotePollBaseTestClass):
+    def create_poll(self) -> None:
+        self.create_model(
+            "poll/1",
+            dict(
+                content_object_id="assignment/1",
+                title="test_title_Zrvh146QAdq7t6iSDwZk",
+                pollmethod="Y",
+                type=Poll.TYPE_NAMED,
+                state=Poll.STATE_CREATED,
+                meeting_id=113,
+                option_ids=[1, 2],
+                entitled_group_ids=[1],
+            ),
+        )
+
+    def test_vote(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": 1, "2": 0}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 200)
+        self.assert_model_exists("vote/1")
+        self.assert_model_not_exists("vote/2")
+        poll = self.get_model("poll/1")
+        self.assertEqual(poll.get("votesvalid"), "1.000000")
+        self.assertEqual(poll.get("votesinvalid"), "0.000000")
+        self.assertEqual(poll.get("votescast"), "1.000000")
+        self.assertEqual(poll.get("state"), Poll.STATE_STARTED)
+        self.assertIn(1, poll.get("voted_ids", []))
+        option1 = self.get_model("option/1")
+        option2 = self.get_model("option/2")
+        self.assertEqual(option1.get("yes"), "1.000000")
+        self.assertEqual(option1.get("no"), "0.000000")
+        self.assertEqual(option1.get("abstain"), "0.000000")
+        self.assertEqual(option2.get("yes"), ".0000000")
+        self.assertEqual(option2.get("no"), "0.000000")
+        self.assertEqual(option2.get("abstain"), "0.000000")
+
+    def test_change_vote(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": 1, "2": 0}, "id": 1, "user_id": 1},
+        )
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": 0, "2": 1}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        option1 = self.get_model("option/1")
+        option2 = self.get_model("option/2")
+        self.assertEqual(option1.get("yes"), "1.000000")
+        self.assertEqual(option1.get("no"), "0.000000")
+        self.assertEqual(option1.get("abstain"), "0.000000")
+        self.assertEqual(option2.get("yes"), "0.000000")
+        self.assertEqual(option2.get("no"), "0.000000")
+        self.assertEqual(option2.get("abstain"), "0.000000")
+
+    def test_global_yes(self) -> None:
+        self.start_poll()
+        response = self.request("poll.vote", {"value": "Y", "id": 1, "user_id": 1})
+        self.assert_status_code(response, 200)
+        poll = self.get_model("poll/1")
+        option = self.get_model("option/1")
+        self.assertEqual(option.get("yes"), "1.000000")
+        self.assertEqual(option.get("no"), "0.000000")
+        self.assertEqual(option.get("abstain"), "0.000000")
+        self.assertEqual(poll.get("amount_global_yes"), "1.000000")
+        self.assertEqual(poll.get("amount_global_no"), "0.000000")
+        self.assertEqual(poll.get("amount_global_abstain"), "0.000000")
+
+    def test_global_yes_forbidden(self) -> None:
+        self.update_model("poll/1", dict(global_yes=False))
+        self.start_poll()
+        response = self.request("poll.vote", {"value": "Y", "id": 1, "user_id": 1})
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+        self.assertEqual(self.get_model("poll/1").get("amount_global_yes"), None)
+
+    def test_global_no(self) -> None:
+        self.start_poll()
+        response = self.request("poll.vote", {"value": "N", "id": 1, "user_id": 1})
+        self.assert_status_code(response, 200)
+        poll = self.get_model("poll/1")
+        option = self.get_model("option/1")
+        self.assertEqual(option.get("yes"), "0.000000")
+        self.assertEqual(option.get("no"), "1.000000")
+        self.assertEqual(option.get("abstain"), "0.000000")
+        self.assertEqual(poll.get("amount_global_yes"), "0.000000")
+        self.assertEqual(poll.get("amount_global_no"), "1.000000")
+        self.assertEqual(poll.get("amount_global_abstain"), "0.000000")
+
+    def test_global_no_forbidden(self) -> None:
+        self.update_model("poll/1", dict(global_no=False))
+        self.start_poll()
+        response = self.request("poll.vote", {"value": "N", "id": 1, "user_id": 1})
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+        self.assertEqual(self.get_model("poll/1").get("amount_global_no"), None)
+
+    def test_global_abstain(self) -> None:
+        self.start_poll()
+        response = self.request("poll.vote", {"value": "A", "id": 1, "user_id": 1})
+        self.assert_status_code(response, 200)
+        poll = self.get_model("poll/1")
+        option = self.get_model("option/1")
+        self.assertEqual(option.get("yes"), "0.000000")
+        self.assertEqual(option.get("no"), "0.000000")
+        self.assertEqual(option.get("abstain"), "1.000000")
+        self.assertEqual(poll.get("amount_global_yes"), "0.000000")
+        self.assertEqual(poll.get("amount_global_no"), "0.000000")
+        self.assertEqual(poll.get("amount_global_abstain"), "1.000000")
+
+    def test_global_abstain_forbidden(self) -> None:
+        self.update_model("poll/1", dict(global_abstain=False))
+        self.start_poll()
+        response = self.request("poll.vote", {"value": "A", "id": 1, "user_id": 1})
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+        self.assertEqual(self.get_model("poll/1").get("amount_global_abstain"), None)
+
+    def test_negative_vote(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": -1}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_too_many_options(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": 1, "2": 1, "3": 1}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_wrong_options(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"3": 1}},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_no_permissions(self) -> None:
+        self.start_poll()
+        # self.make_admin_delegate()
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": 1}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_anonymous(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": 1}, "id": 1, "user_id": 0},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_vote_not_present(self) -> None:
+        self.start_poll()
+        self.update_model("user/1", dict(is_present_in_meeting_ids=[]))
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": 1}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_wrong_state(self) -> None:
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": 1}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_missing_data(self) -> None:
+        self.start_poll()
+        response = self.request("poll.vote", {"value": {}, "id": 1, "user_id": 1})
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+        poll = self.get_model("poll/1")
+        self.assertNotIn(1, poll.get("voted_ids", []))
+
+    def test_wrong_data_format(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": [1, 2, 5], "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_wrong_option_format(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": "string"}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_wrong_option_id_type(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"id": 1}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_wrong_vote_data(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": [None]}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+
+class VotePollOS3NamedN(VotePollBaseTestClass):
+    def create_poll(self) -> None:
+        self.create_model(
+            "poll/1",
+            dict(
+                content_object_id="assignment/1",
+                title="test_title_4oi49ckKFk39SDIfj30s",
+                pollmethod="N",
+                type=Poll.TYPE_NAMED,
+                state=Poll.STATE_CREATED,
+                meeting_id=113,
+                option_ids=[1, 2],
+                entitled_group_ids=[1],
+            ),
+        )
+
+    def test_vote(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": 1, "2": 0}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 200)
+        self.assert_model_exists("vote/1")
+        self.assert_model_not_exists("vote/2")
+        poll = self.get_model("poll/1")
+        self.assertEqual(poll.get("votesvalid"), "1.000000")
+        self.assertEqual(poll.get("votesinvalid"), "0.000000")
+        self.assertEqual(poll.get("votescast"), "1.000000")
+        self.assertEqual(poll.get("state"), Poll.STATE_STARTED)
+        self.assertTrue(1 in poll.get("voted_ids", []))
+        option1 = self.get_model("option/1")
+        option2 = self.get_model("option/2")
+        self.assertEqual(option1.get("yes"), "0.000000")
+        self.assertEqual(option1.get("no"), "1.000000")
+        self.assertEqual(option1.get("abstain"), "0.000000")
+        self.assertEqual(option2.get("yes"), "0.000000")
+        self.assertEqual(option2.get("no"), "0.000000")
+        self.assertEqual(option2.get("abstain"), "0.000000")
+
+    def test_change_vote(self) -> None:
+        self.add_candidate()
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": 1, "2": 0}, "id": 1, "user_id": 1},
+        )
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": 0, "2": 1}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        option1 = self.get_model("option/1")
+        option2 = self.get_model("option/2")
+        self.assertEqual(option1.get("yes"), "0.000000")
+        self.assertEqual(option1.get("no"), "1.000000")
+        self.assertEqual(option1.get("abstain"), "0.000000")
+        self.assertEqual(option2.get("yes"), "0.000000")
+        self.assertEqual(option2.get("no"), "0.000000")
+        self.assertEqual(option2.get("abstain"), "0.000000")
+
+    def test_global_yes(self) -> None:
+        self.start_poll()
+        response = self.request("poll.vote", {"value": "Y", "id": 1, "user_id": 1})
+        self.assert_status_code(response, 200)
+        poll = self.get_model("poll/1")
+        option = self.get_model("option/1")
+        self.assertEqual(option.get("yes"), "1")
+        self.assertEqual(option.get("no"), "0")
+        self.assertEqual(option.get("abstain"), "0")
+        self.assertEqual(poll.get("amount_global_yes"), "1")
+        self.assertEqual(poll.get("amount_global_no"), "0")
+        self.assertEqual(poll.get("amount_global_abstain"), "0")
+
+    def test_global_yes_forbidden(self) -> None:
+        self.update_model("poll/1", dict(global_yes=False))
+        self.start_poll()
+        response = self.request("poll.vote", {"value": "Y", "id": 1, "user_id": 1})
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+        poll = self.get_model("poll/1")
+        self.assertEqual(poll.get("amount_global_yes"), None)
+
+    def test_global_no(self) -> None:
+        self.start_poll()
+        response = self.request("poll.vote", {"value": "N", "id": 1, "user_id": 1})
+        self.assert_status_code(response, 200)
+        poll = self.get_model("poll/1")
+        option = self.get_model("option/1")
+        self.assertEqual(option.get("yes"), "0.000000")
+        self.assertEqual(option.get("no"), "1.000000")
+        self.assertEqual(option.get("abstain"), "0.000000")
+        self.assertEqual(poll.get("amount_global_yes"), "0.000000")
+        self.assertEqual(poll.get("amount_global_no"), "1.000000")
+        self.assertEqual(poll.get("amount_global_abstain"), "0.000000")
+
+    def test_global_no_forbidden(self) -> None:
+        self.update_model("poll/1", dict(global_no=False))
+        self.start_poll()
+        response = self.request("poll.vote", {"value": "N", "id": 1, "user_id": 1})
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+        self.assertEqual(self.get_model("poll/1").get("amount_global_no"), None)
+
+    def test_global_abstain(self) -> None:
+        self.start_poll()
+        response = self.request("poll.vote", {"value": "A", "id": 1, "user_id": 1})
+        self.assert_status_code(response, 200)
+        poll = self.get_model("poll/1")
+        option = self.get_model("option/1")
+        self.assertEqual(option.get("yes"), "0.000000")
+        self.assertEqual(option.get("no"), "0.000000")
+        self.assertEqual(option.get("abstain"), "1.000000")
+        self.assertEqual(poll.get("amount_global_yes"), "0.000000")
+        self.assertEqual(poll.get("amount_global_no"), "0.000000")
+        self.assertEqual(poll.get("amount_global_abstain"), "1.000000")
+
+    def test_global_abstain_forbidden(self) -> None:
+        self.update_model("poll/1", dict(global_abstain=False))
+        self.start_poll()
+        response = self.request("poll.vote", {"value": "A", "id": 1, "user_id": 1})
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+        self.assertEqual(self.get_model("poll/1").get("amount_global_abstain"), None)
+
+    def test_negative_vote(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": -1}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_wrong_options(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"3": 1}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_no_permissions(self) -> None:
+        self.start_poll()
+        # self.make_admin_delegate()
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": 1}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_anonymous(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": 1}, "id": 1, "user_id": 2},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_vote_not_present(self) -> None:
+        self.start_poll()
+        self.update_model("user/1", dict(is_present_in_meeting_ids=[]))
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": 1}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_wrong_state(self) -> None:
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": 1}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_missing_data(self) -> None:
+        self.start_poll()
+        response = self.request("poll.vote", {"value": {}, "id": 1, "user_id": 1})
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+        poll = self.get_model("poll/1")
+        self.assertNotIn(1, poll.get("voted_ids", []))
+
+    def test_wrong_data_format(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": [1, 2, 5], "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_wrong_option_format(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": "string"}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_wrong_option_id_type(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"id": 1}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_wrong_vote_data(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": [None]}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+
+class VotePollOS3PseudoanonymousYNA(VotePollBaseTestClass):
+    def create_poll(self) -> None:
+        self.create_model(
+            "poll/1",
+            dict(
+                content_object_id="assignment/1",
+                title="test_title_OkHAIvOSIcpFnCxbaL6v",
+                pollmethod="YNA",
+                type=Poll.TYPE_PSEUDOANONYMOUS,
+                state=Poll.STATE_CREATED,
+                meeting_id=113,
+                option_ids=[1, 2],
+                entitled_group_ids=[1],
+            ),
+        )
+
+    def test_vote(self) -> None:
+        self.add_candidate()
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": "Y", "2": "N", "3": "A"}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 200)
+        self.get_model("vote/1")
+        self.get_model("vote/2")
+        self.get_model("vote/3")
+        self.assert_model_not_exists("vote/4")
+        poll = self.get_model("poll/1")
+        self.assertEqual(poll.get("votesvalid"), "1.000000")
+        self.assertEqual(poll.get("votesinvalid"), "0.000000")
+        self.assertEqual(poll.get("votescast"), "1.000000")
+        self.assertEqual(poll.get("state"), Poll.STATE_STARTED)
+        option1 = self.get_model("option/1")
+        option2 = self.get_model("option/2")
+        option3 = self.get_model("option/3")
+        self.assertEqual(option1.get("yes"), "1.000000")
+        self.assertEqual(option1.get("no"), "0.000000")
+        self.assertEqual(option1.get("abstain"), "0.000000")
+        self.assertEqual(option2.get("yes"), "0.000000")
+        self.assertEqual(option2.get("no"), "1.000000")
+        self.assertEqual(option2.get("abstain"), "0.000000")
+        self.assertEqual(option3.get("yes"), "0.000000")
+        self.assertEqual(option3.get("no"), "0.000000")
+        self.assertEqual(option3.get("abstain"), "1.000000")
+
+    def test_change_vote(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": "Y"}, "id": 1, "user_id": 1},
+        )
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": "N"}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        option1 = self.get_model("option/1")
+        self.assertEqual(option1.get("yes"), "1.000000")
+        self.assertEqual(option1.get("no"), "0.000000")
+        self.assertEqual(option1.get("abstain"), "0.000000")
+
+    def test_too_many_options(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": "Y", "3": "N"}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_partial_vote(self) -> None:
+        self.add_candidate()
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": "Y"}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 200)
+        self.assert_model_exists("vote/1")
+
+    def test_wrong_options(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": "Y", "3": "N"}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_no_permissions(self) -> None:
+        self.start_poll()
+        # self.make_admin_delegate()
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": "Y"}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_anonymous(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": "Y"}, "id": 1, "user_id": 2},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_vote_not_present(self) -> None:
+        self.start_poll()
+        self.update_model("user/1", dict(is_present_in_meeting_ids=[]))
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": "Y"}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_wrong_state(self) -> None:
+        response = self.request("poll.vote", {"value": {}, "id": 1, "user_id": 1})
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_missing_value(self) -> None:
+        self.start_poll()
+        response = self.request("poll.vote", {"value": {}, "id": 1, "user_id": 1})
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+        poll = self.get_model("poll/1")
+        self.assertNotIn(1, poll.get("voted_ids", []))
+
+    def test_wrong_value_format(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": [1, 2, 5], "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_wrong_option_format(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": "string"}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_wrong_option_id_type(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"id": "Y"}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_wrong_vote_value(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": [None]}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+
+class VotePollOS3PseudoanonymousY(VotePollBaseTestClass):
+    def create_poll(self) -> None:
+        self.create_model(
+            "poll/1",
+            dict(
+                content_object_id="assignment/1",
+                title="test_title_Zrvh146QAdq7t6iSDwZk",
+                pollmethod="Y",
+                type=Poll.TYPE_PSEUDOANONYMOUS,
+                state=Poll.STATE_CREATED,
+                meeting_id=113,
+                option_ids=[1, 2],
+                entitled_group_ids=[1],
+            ),
+        )
+
+    def test_vote(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": 1, "2": 0}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 200)
+        self.assert_model_exists("vote/1")
+        self.assert_model_not_exists("vote/2")
+        poll = self.get_model("poll/1")
+        self.assertEqual(poll.get("votesvalid"), "1.000000")
+        self.assertEqual(poll.get("votesinvalid"), "0.000000")
+        self.assertEqual(poll.get("votescast"), "1.000000")
+        self.assertEqual(poll.get("state"), Poll.STATE_STARTED)
+        self.assertTrue(1 in poll.get("voted_ids", []))
+        option1 = self.get_model("option/1")
+        option2 = self.get_model("option/2")
+        self.assertEqual(option1.get("yes"), "1.000000")
+        self.assertEqual(option1.get("no"), "0.000000")
+        self.assertEqual(option1.get("abstain"), "0.000000")
+        self.assertEqual(option2.get("yes"), "0.000000")
+        self.assertEqual(option2.get("no"), "0.000000")
+        self.assertEqual(option2.get("abstain"), "0.000000")
+        vote = self.get_model("vote/1")
+        self.assertIsNone(vote.get("user_id"))
+
+    def test_change_vote(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": 1, "2": 0}, "id": 1, "user_id": 1},
+        )
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": 0, "2": 1}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.get_model("poll/1")
+        option1 = self.get_model("option/1")
+        option2 = self.get_model("option/2")
+        self.assertEqual(option1.get("yes"), "1.000000")
+        self.assertEqual(option1.get("no"), "0.000000")
+        self.assertEqual(option1.get("abstain"), "0.000000")
+        self.assertEqual(option2.get("yes"), "0.000000")
+        self.assertEqual(option2.get("no"), "0.000000")
+        self.assertEqual(option2.get("abstain"), "0.000000")
+
+    def test_negative_vote(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": -1}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_wrong_options(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"3": 1}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_vote_not_present(self) -> None:
+        self.start_poll()
+        self.update_model("user/1", dict(is_present_in_meeting_ids=[]))
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": 1}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_wrong_state(self) -> None:
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": 1}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_missing_data(self) -> None:
+        self.start_poll()
+        response = self.request("poll.vote", {"value": {}, "id": 1, "user_id": 1})
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+        poll = self.get_model("poll/1")
+        self.assertNotIn(1, poll.get("voted_ids", []))
+
+    def test_wrong_data_format(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"value": [1, 2, 5]}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_wrong_option_format(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": "string"}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_wrong_option_id_type(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"id": 1}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_wrong_vote_data(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": [None]}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+
+class VotePollOS3PseudoAnonymousN(VotePollBaseTestClass):
+    def create_poll(self) -> None:
+        self.create_model(
+            "poll/1",
+            dict(
+                content_object_id="assignment/1",
+                title="test_title_wWPOVJgL9afm83eamf3e",
+                pollmethod="N",
+                type=Poll.TYPE_PSEUDOANONYMOUS,
+                state=Poll.STATE_CREATED,
+                meeting_id=113,
+                option_ids=[1, 2],
+                entitled_group_ids=[1],
+            ),
+        )
+
+    def test_vote(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"id": 1, "value": {"1": 1, "2": 0}, "user_id": 1},
+        )
+        self.assert_status_code(response, 200)
+        self.assert_model_exists("vote/1")
+        self.assert_model_not_exists("vote/2")
+        poll = self.get_model("poll/1")
+        self.assertEqual(poll.get("votesvalid"), "1.000000")
+        self.assertEqual(poll.get("votesinvalid"), "0.000000")
+        self.assertEqual(poll.get("votescast"), "1.000000")
+        self.assertEqual(poll.get("state"), Poll.STATE_STARTED)
+        self.assertTrue(1 in poll.get("voted_ids", []))
+        option1 = self.get_model("option/1")
+        option2 = self.get_model("option/2")
+        self.assertEqual(option1.get("yes"), "0")
+        self.assertEqual(option1.get("no"), "1")
+        self.assertEqual(option1.get("abstain"), "0")
+        self.assertEqual(option2.get("yes"), "0")
+        self.assertEqual(option2.get("no"), "0")
+        self.assertEqual(option2.get("abstain"), "0")
+        vote = self.get_model("vote/1")
+        self.assertIsNone(vote.get("user_id"))
+
+    def test_change_vote(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": 1, "2": 0}, "id": 1, "user_id": 1},
+        )
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": 0, "2": 1}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.get_model("poll/1")
+        option1 = self.get_model("option/1")
+        option2 = self.get_model("option/2")
+        self.assertEqual(option1.get("yes"), "0.000000")
+        self.assertEqual(option1.get("no"), "1.000000")
+        self.assertEqual(option1.get("abstain"), "0.000000")
+        self.assertEqual(option2.get("yes"), "0.000000")
+        self.assertEqual(option2.get("no"), "0.000000")
+        self.assertEqual(option2.get("abstain"), "0.000000")
+
+    def test_negative_vote(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": -1}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_vote_not_present(self) -> None:
+        self.start_poll()
+        self.update_model("user/1", dict(is_present_in_meeting_ids=[]))
+
+        response = self.request(
+            "poll.vote",
+            {"id": 1, "user_id": 1, "value": {"1": 1}},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_wrong_state(self) -> None:
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": 1}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_missing_data(self) -> None:
+        self.start_poll()
+        response = self.request("poll.vote", {})
+        self.assert_status_code(response, 400)
+        assert (
+            "data must contain ['id', 'user_id', 'value'] properties"
+            in response.data.decode()
+        )
+        self.assert_model_not_exists("vote/1")
+        poll = self.get_model("poll/1")
+        self.assertNotIn(1, poll.get("voted_ids", []))
+
+    def test_wrong_data_format(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": [1, 2, 5], "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        assert "data.value must be object or string" in response.data.decode()
+        self.assert_model_not_exists("vote/1")
+
+    def test_wrong_option_format(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"value": {"1": "string"}, "id": 1, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        assert "Option 1 has not a right value. (int, str)." in response.data.decode()
+        self.assert_model_not_exists("vote/1")
+
+    def test_wrong_option_id_type(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"id": 1},
+        )
+        self.assert_status_code(response, 400)
+        self.assert_model_not_exists("vote/1")
+
+    def test_wrong_vote_data(self) -> None:
+        self.start_poll()
+        response = self.request(
+            "poll.vote",
+            {"id": 1, "value": {"1": [None]}, "user_id": 1},
+        )
+        self.assert_status_code(response, 400)
+        assert "Option 1 has not a right value. (int, str)." in response.data.decode()
+        self.assert_model_not_exists("vote/1")
